semantic-search/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.0
pydantic-settings==2.6.0

# Vector database
qdrant-client==1.12.1

# ML and embeddings
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.46.3

# Data processing
numpy==2.2.1
pandas==2.2.3

# HTTP and async
httpx==0.28.1
aiofiles==24.1.0

# Utilities
python-dotenv==1.0.1
redis==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging and monitoring
structlog==24.4.0
